PLEASE READ
I did alot here, So I added basic Menu shit just for placement and a few sprites, I will have more work done to that, specifically to the appearance, I'll work on the code for that. I did a SHIT ton of organization with the sprites and objects so I recommend checking that out. If none of the changes load, PELASE DO NOT DO WORK ON ANYTHING.